Firmware/Tool: 16 bit colors + better gamma table

+++ firmware/ws2812b.h
@@ -1,6 +1,6 @@
 #ifndef _WS2812B_H
 #define _WS2812B_H
 
-void ws2812b_set_rgb(uint8_t r, uint8_t g, uint8_t b);
+void ws2812b_set_rgb(uint16_t r, uint16_t g, uint16_t b);
 
 #endif
